move prefix logic back to testutil
