+++ internal/testutil/backend.go
@@ -268,7 +268,7 @@ func (b *MemoryBackend) ListObjects(ctx context.Context, accessKeyID *string, bu
 
 	var lastMatchedPart string
 	for _, obj := range objects {
-		match := s3.Match(prefix, obj.name)
+		match := match(prefix, obj.name)
 		switch {
 		case match == nil:
 			continue
